requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-groq>=0.0.1
pandas>=2.0.0
chromadb>=0.4.22
streamlit>=1.32.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
openpyxl>=3.1.2
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0 